Update check-username-availability.php
I would add or die(mysql_error()); to the end of the sql line. It will display any errors in your query in the first place. It would be like. Below:

 $username_query = mysql_query("SELECT COUNT(user_id) FROM users WHERE username = '$username'") or die(mysql_error());

Also COUNT has to be like COUNT(column_name)

Also the mysql_ functions are depecrated. It would be wise to use mysqli or PDO